cooking_chatbot/requirements.txt
streamlit==1.32.0
fastapi==0.109.2
uvicorn==0.27.1
anthropic==0.18.1
spoonacular==3.0
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
sqlalchemy==2.0.27